+++ get_oauth_token.php
@@ -5,18 +5,14 @@
  * as [https/http]://<yourdomain>/<folder>/get_oauth_token.php
  * e.g.: http://localhost/phpmail/get_oauth_token.php
  * * Ensure dependencies are installed with 'composer install'
- * * Set up an app in your Google/Yahoo/Microsoft
+ * * Set up an app in your Google/Yahoo/Microsoft account
  * * Set the script address as the app's redirect URL
  * If no refresh token is obtained when running this file, revoke access to your app
  * using link: https://accounts.google.com/b/0/IssuedAuthSubTokens and run the script again.
  * This script requires PHP 5.4 or later
- * PHP Version 5.4.
+ * PHP Version 5.4
  */
 require 'vendor/autoload.php';
-require 'class.phpmaileroauthprovider.php';
-require 'class.phpmaileroauthgoogle.php';
-require 'class.phpmaileroauthyahoo.php';
-require 'class.phpmaileroauthmicrosoft.php';
 
 session_start();
 
@@ -52,7 +48,7 @@ if ($provider_name == 'google') {
     $options['scope'] = array(
         'https://mail.google.com/'
     );
-    
+
     //To get the refresh token everytime
     $options['approval_prompt'] = 'force';
 } elseif ($provider_name == 'yahoo') {
@@ -67,7 +63,7 @@ if ($provider_name == 'google') {
         'clientSecret' => '{MICROSOFT_APP_SECRET}',
         'redirectUri' => $redirectUri
     ));
-    
+
     //scopes
     $options['scope'] = array(
         'wl.imap',
@@ -79,34 +75,33 @@ if ($provider_name == 'google') {
 }
 
 if (!isset($_GET['code'])) {
-    
     // If we don't have an authorization code then get one
     $authUrl = $provider->getAuthorizationUrl($options);
-    
+
     // echo $authUrl."<br>";
     $_SESSION['oauth2state'] = $provider->getState();
-    
+
     header('Location: ' . $authUrl);
     exit;
-    // Check given state against previously stored one to mitigate CSRF attack
+// Check given state against previously stored one to mitigate CSRF attack
 } elseif (empty($_GET['state']) || ($_GET['state'] !== $_SESSION['oauth2state'])) {
     unset($_SESSION['oauth2state']);
     unset($_SESSION['provider_name']);
     exit('Invalid state');
 } else {
     unset($_SESSION['provider_name']);
-    
+
     // Try to get an access token (using the authorization code grant)
     $token = $provider->getAccessToken('authorization_code', array(
         'code' => $_GET['code']
     ));
-    
+
     // Use this to interact with an API on the users behalf
     echo 'Token: ' . $token->getToken() . '<br>';
-    
+
     // Use this to get a new access token if the old one expires
     echo 'Refresh Token: ' . $token->getRefreshToken() . '<br>';
-    
+
     // Number of seconds until the access token will expire, and need refreshing
     echo 'Expires:' . $token->getExpires() . '<br>';
 }